refactor(selecteur): les deux autocomplétions sortent d'app.js, et la carte du vaisseau cesse d'être un `let` vide
`montePicker`, le sélecteur de station (ADR-003) et celui des vaisseaux emménagent dans
`selecteur.js` — 288 lignes, et `app.js` passe de 1 549 à 1 334.

Ils restent IMPÉRATIFS, et c'est la règle de la frontière (ADR-012 §2) : `#stationPickList` et
`#shipList` sont du markup d'`index.html` qu'aucun portail ne possède, et `monterAutocompletion`
sert deux widgets aux contrats différents — l'un groupé par système › zone et non plafonné, l'autre
plat et coupé à 12. Les réunir derrière une douzaine de props serait la migration de syntaxe que
cette refonte refuse.

Un vrai défaut corrigé au passage : `showShipCard` était un `let` de la coquille RÉASSIGNÉ depuis
l'intérieur de `loadShips`. Lu avant la réponse du fetch, il appelait une fonction vide — sans
erreur, sans trace. La liste des vaisseaux est maintenant l'état du module, et
`montrerCarteVaisseau()` la lit.

Parti avec eux : le garde du champ `#station` (`derniereStation`), qui n'a de sens qu'à côté de ce
qui l'écrit, et qui se lit désormais `if (stationChangee()) refresh();`. `esc` monte dans
`format.ts` — une seule source pour ses quatre appelants restants.
